[PT-1240] refactor: saengbu → saenggibu 명칭 정정
생기부 로마자를 정확히(생·기·부) — 핸들러 파일·테스트·dispatch import 를 saenggibu 로
통일. 스트림 키(myfolio:stream:saenggibu-*)·규약 문서와 일관.
